+++ src/components/OverflowList.tsx
@@ -1,4 +1,4 @@
-import React, { useRef, useState, Activity } from "react";
+import React, { useRef, useState } from "react";
 import { useForkRef, useIsoLayoutEffect, useResizeObserver } from "../hooks";
 import { getRowPositionsData } from "../utils";
 import { DefaultOverflowElement } from "./DefaultOverflowMenu";
@@ -32,11 +32,8 @@ type BaseOverflowListProps<T> = BaseComponentProps & {
   // if false, using requestAnimationFrame to update the state - this avoid forced reflow and improve performance
   flushImmediately?: boolean;
 
-  // use this if you want to override how hidden items are kept in the DOM while not affecting layout
-  // this function is called for item item in the list(even visible ones), you can control the visibility of the item by returning a different node or null
-  //   default: about react 19.2 new Activity component to control the visibility of the item while don't forcing mount/unmount of the item
-  //            below react 19.2 is simply null (causing rapid re-mount/unmount of the item)
-  renderHiddenItem?: (node: React.ReactNode, meta: RenderItemMeta) => React.ReactNode;
+  // customize how each item is shown/hidden during measurement so you can keep custom elements mounted
+  renderItemVisibility?: (node: React.ReactNode, meta: RenderItemMeta) => React.ReactNode;
 };
 
 type OverflowListWithItems<T> = BaseOverflowListProps<T> & {
@@ -86,7 +83,7 @@ const OverflowListComponent = React.memo(
       renderItem = (item) => item as React.ReactNode,
       renderOverflowItem,
       renderOverflowProps,
-      renderHiddenItem,
+      renderItemVisibility,
       maxRows = 1,
       maxVisibleItems = 100,
       flushImmediately = true,
@@ -224,8 +221,8 @@ const OverflowListComponent = React.memo(
       ...containerProps.style,
     };
 
-    const finalRenderHiddenItem =
-      renderHiddenItem ??
+    const finalRenderItemVisibility =
+      renderItemVisibility ??
       ((node, meta) => {
         // prefer react 19.2 new activity component to control the visibility of the item while don't forcing mount/unmount of the item
         const Activity = React?.Activity;
@@ -254,7 +251,7 @@ const OverflowListComponent = React.memo(
 
           const itemComponent = renderItem(item as T, { index, visible: isVisible });
 
-          return finalRenderHiddenItem(itemComponent, { index, visible: isVisible });
+          return finalRenderItemVisibility(itemComponent, { index, visible: isVisible });
         })}
 
         {clonedOverflowElement}